accidental jwt inclusion, but refactoring react
