+++ react/hello-world/src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import ContainerTable from './components/containers';
-import App from './components/loginform';
+//import ContainerTable from './components/ContainerTable';
+import App from './components/App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
